added maptiles_utils file
